Time: 44 ms (21.7%), Space: 68.6 MB (90.8%) - LeetHub

+++ 1814-count-nice-pairs-in-an-array/1814-count-nice-pairs-in-an-array.java
@@ -23,9 +23,9 @@ class Solution {
                 map.put(a,map.get(a) + 1);
             }
             else{
-                map.put(a, 1l);
+                map.put(a, 1);
             }
         }
-        return count;
+        return (int)count;
     }
 }
